Refactor `CSVLogger` to use `LogOutput` class.
Now, the `CSVLogger` receives a `LogOutput` instance in its constructor
that determines the type of output it wants to write the logs into. This
could be helpful for unit testing the new loggers to make sure their
output looks as expected.
