+++ units/csv_log_tests.cc
@@ -12,16 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <cstdio>
+#include <iostream>
+#include <vector>
+
 #include "csv_logging.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
+
+using ::testing::ElementsAre;
 
 namespace performancelayers {
 namespace {
 // This is a simple test and only calls the methods to make sure the
 // logger doesn't crash.
 TEST(CSVLogger, MethodCheck) {
-  CSVLogger logger("pipeline,duration", nullptr);
+  StringOutput out = {};
+  CSVLogger logger("pipeline,duration", &out);
   VectorInt64Attr hashes("hashes", {2, 3});
   Duration dur = Duration::FromNanoseconds(1);
   CreateGraphicsPipelinesEvent pipeline_event("create_graphics_pipeline",
@@ -32,8 +39,8 @@ TEST(CSVLogger, MethodCheck) {
   logger.EndLog();
   // Checks double `EndLog` calls.
   logger.EndLog();
-
-  SUCCEED();
+  std::string_view event_str = "\"[0x2,0x3]\",1";
+  EXPECT_THAT(out.GetLog(), ElementsAre("pipeline,duration", event_str));
 }
 
 }  // namespace
